Apply master key

+++ src/app/controller/CTRAdjectives.js
@@ -1,4 +1,6 @@
 const DBAdjectives = require('../../db/DBAdjectives')
+const config = require('../../config/db.connect.config');
+
 
 async function selectAll(req, res) {
     var result = await DBAdjectives.selectAll()
@@ -6,22 +8,34 @@ async function selectAll(req, res) {
 }
 
 async function insertByName(req, res) {
-    var result = await DBAdjectives.insertByName(req.body.name)
-    if(result){
-        res.sendStatus(200)
+
+    if(req.body.masterKey==config.password){
+        var result = await DBAdjectives.insertByName(req.body.name)
+        if(result){
+            res.send('200')
+        }
+        else{
+            res.send(`${req.body.name} is already exist`)        
+        }
     }
     else{
-        res.send(`${req.body.name} is already exist`)        
+        res.send(`password is wrong`)  
     }
 }
 
 async function deleteByName(req, res) {
-    var result = await DBAdjectives.deleteByName(req.body.name)
-    if(result){
-        res.sendStatus(200)
+
+    if(req.body.masterKey==config.password){
+        var result = await DBAdjectives.deleteByName(req.body.name)
+        if(result){
+            res.send('200')
+        }
+        else{
+            res.send(`${req.body.name} is not exist`)        
+        }
     }
     else{
-        res.send(`${req.body.name} is not exist`)        
+        res.send(`password is wrong`)  
     }
 }
